+++ frontend/src/components/ui/MapController.tsx
@@ -7,7 +7,7 @@ import { useCurrentPosition } from '@hooks/google-maps/useCurrentPosition';
 import Box from '@common/Box';
 import ButtonNext from '@common/ButtonNext';
 
-import { INITIAL_ZOOM_SIZE } from '@constants';
+import { INITIAL_ZOOM_SIZE } from '@constants/googleMaps';
 
 const MapController = () => {
   const position = useCurrentPosition();
