+++ frontend/src/constants/googleMaps.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_CENTER = {
+  lat: 37.5056102333107,
+  lng: 127.05081496722168,
+} as const;
+export const INITIAL_ZOOM_SIZE = 15;
