+++ frontend/src/hooks/tanstack-query/station-markers/useStations.ts
@@ -17,7 +17,8 @@ import {
 import { serverStore } from '@stores/serverStore';
 import { stationFilterStore } from '@stores/stationFilterStore';
 
-import { INITIAL_ZOOM_SIZE, SERVERS } from '@constants';
+import { SERVERS } from '@constants';
+import { INITIAL_ZOOM_SIZE } from '@constants/googleMaps';
 
 import type { DisplayPosition, StationSummary } from '../../../types';
